add Navbar component
