+++ src/App.tsx
@@ -1,31 +1,11 @@
 import './App.scss';
-import cn from 'classnames';
-import { NavLink, Outlet } from 'react-router-dom';
-
-const getLinkClass = ({ isActive }: { isActive: boolean }) =>
-  cn('navbar-item', { 'has-background-grey-lighter': isActive });
+import { Outlet } from 'react-router-dom';
+import { Navbar } from './components/Navbar';
 
 export const App = () => {
   return (
     <div data-cy="app">
-      <nav
-        data-cy="nav"
-        className="navbar is-fixed-top has-shadow"
-        role="navigation"
-        aria-label="main navigation"
-      >
-        <div className="container">
-          <div className="navbar-brand">
-            <NavLink to="/" className={getLinkClass}>
-              Home
-            </NavLink>
-
-            <NavLink to="/people" className={getLinkClass}>
-              People
-            </NavLink>
-          </div>
-        </div>
-      </nav>
+      <Navbar />
 
       <main className="section">
         <div className="container">
